+++ modules/shared/mcp/src/sap/commerce/toolset/shared/mcp/transform/dto/TransformerInfo.kt
@@ -16,11 +16,13 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sap.commerce.toolset.shared.mcp.dto
+package sap.commerce.toolset.shared.mcp.transform.dto
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class TransformerMcpResult(
-    val languages: List<LanguageTransformers>
+data class TransformerInfo(
+    val id: String,
+    val name: String,
+    val description: String,
 )
